pyproject.toml
[tool.poetry]
name = "pyforms"
version = "0.1.0"
description = ""
authors = ["Roberto Polli <robipolli@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
reportlab = "^3.5.42"
pyyaml = "^5.3.1"
jsonschema = "^3.2.0"
pypdf = "^1.13"
pypdf2 = "^1.26.0"
jsonref = "^0.2"

[tool.poetry.dev-dependencies]
ipython = "^7.14.0"
pytest = "^5.4.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
